more checkin

+++ cpp/reverse_integer.cpp
@@ -32,7 +32,7 @@ public:
             x=-x;
         }
         
-        unsigned long long res=0;
+        unsigned long res=0;
         while (x>0) {            
             res=res*10+x%10;
             x/=10;
